temp

+++ test/graphplot.py
@@ -20,7 +20,7 @@ for x in xpoints:
 
     out = out.decode()
     value = out.split('\n')[1].split()[4:6]
-    value[0] = int(value[0])
+    value[0] = float(value[0])
     if (value[1] != 'ms'):
         value[0]  *= 1000
     ypoints.append(value[0])
@@ -56,7 +56,7 @@ for x in xpoints:
 
     out = out.decode()
     value = out.split('\n')[1].split()[4:6]
-    value[0] = int(value[0])
+    value[0] = float(value[0])
     if (value[1] != 'ms'):
         value[0]  *= 1000
     ypoints.append(value[0])
